Se realizaron ajustes menores para mejorar el funcionamiento de cada página de la aplicación web. Además, se agregaron mensajes de confirmación al eliminar y registrar usuarios. Queda pendiente la validación y redirección de datos y las mejoras opcionales en la interfaz utilizando CSS y Bootstrap. [25/02/24 09:56]

+++ app/Http/Controllers/PeliculaController.php
@@ -5,7 +5,8 @@ namespace App\Http\Controllers;
 use App\Models\Pelicula;
 use Illuminate\Http\Request;
 
-use Illuminate\Support\Facades\Storage; 
+use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\DB;
 
 class PeliculaController extends Controller
 {
@@ -47,7 +48,8 @@ class PeliculaController extends Controller
             $datosPelicula['foto']=$request->file('foto')->store('uploads','public');
         }
         Pelicula::insert($datosPelicula);
-         return response()->json($datosPelicula);
+        //return response()->json($datosPelicula);
+        return redirect('pelicula')->with('mensaje','Pelicula agregada con exito');
     }
 
     /**
@@ -105,8 +107,14 @@ class PeliculaController extends Controller
      */
     public function destroy($id)
     {
-        //
+    //
+    $pelicula = Pelicula::findOrFail($id);
+
+    if (Storage::delete('public/'.$pelicula->foto)){
         Pelicula::destroy($id);
-        return redirect('pelicula');
+    }
+
+
+    return redirect('pelicula')->with('mensaje','Pelicula eliminada');
     }
 }
